[BEAM-6725] share some Flink job invocation code with Samza runner
