+++ runners/flink/src/test/java/org/apache/beam/runners/flink/PortableExecutionTest.java
@@ -26,6 +26,7 @@ import org.apache.beam.model.pipeline.v1.RunnerApi;
 import org.apache.beam.runners.core.construction.Environments;
 import org.apache.beam.runners.core.construction.JavaReadViaImpulse;
 import org.apache.beam.runners.core.construction.PipelineTranslation;
+import org.apache.beam.runners.fnexecution.jobsubmission.JobInvocation;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.coders.BigEndianLongCoder;
 import org.apache.beam.sdk.coders.KvCoder;
@@ -140,8 +141,8 @@ public class PortableExecutionTest implements Serializable {
     RunnerApi.Pipeline pipelineProto = PipelineTranslation.toProto(p);
 
     // execute the pipeline
-    FlinkJobInvocation jobInvocation =
-        FlinkJobInvocation.create(
+    JobInvocation jobInvocation =
+        FlinkJobInvoker.createJobInvocation(
             "fakeId",
             "fakeRetrievalToken",
             flinkJobExecutor,
