+++ runners/flink/src/test/java/org/apache/beam/runners/flink/PortableStateExecutionTest.java
@@ -26,6 +26,7 @@ import org.apache.beam.model.pipeline.v1.RunnerApi;
 import org.apache.beam.runners.core.construction.Environments;
 import org.apache.beam.runners.core.construction.JavaReadViaImpulse;
 import org.apache.beam.runners.core.construction.PipelineTranslation;
+import org.apache.beam.runners.fnexecution.jobsubmission.JobInvocation;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.coders.VarIntCoder;
 import org.apache.beam.sdk.options.PipelineOptions;
@@ -194,8 +195,8 @@ public class PortableStateExecutionTest implements Serializable {
 
     RunnerApi.Pipeline pipelineProto = PipelineTranslation.toProto(p);
 
-    FlinkJobInvocation jobInvocation =
-        FlinkJobInvocation.create(
+    JobInvocation jobInvocation =
+        FlinkJobInvoker.createJobInvocation(
             "id",
             "none",
             flinkJobExecutor,
