+++ runners/java-fn-execution/src/main/java/org/apache/beam/runners/fnexecution/jobsubmission/JobInvoker.java
@@ -18,13 +18,40 @@
 package org.apache.beam.runners.fnexecution.jobsubmission;
 
 import java.io.IOException;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ThreadFactory;
 import javax.annotation.Nullable;
 import org.apache.beam.model.pipeline.v1.RunnerApi;
 import org.apache.beam.vendor.grpc.v1p13p1.com.google.protobuf.Struct;
+import org.apache.beam.vendor.guava.v20_0.com.google.common.util.concurrent.ListeningExecutorService;
+import org.apache.beam.vendor.guava.v20_0.com.google.common.util.concurrent.MoreExecutors;
+import org.apache.beam.vendor.guava.v20_0.com.google.common.util.concurrent.ThreadFactoryBuilder;
 
-/** Factory to create a {@link JobInvocation} instances. */
-public interface JobInvoker {
+/** Factory to create {@link JobInvocation} instances. */
+public abstract class JobInvoker {
+
+  private final ListeningExecutorService executorService;
+
   /** Start running a job, abstracting its state as a {@link JobInvocation} instance. */
-  JobInvocation invoke(RunnerApi.Pipeline pipeline, Struct options, @Nullable String retrievalToken)
+  protected abstract JobInvocation invokeWithExecutor(
+      RunnerApi.Pipeline pipeline,
+      Struct options,
+      @Nullable String retrievalToken,
+      ListeningExecutorService executorService)
       throws IOException;
+
+  JobInvocation invoke(RunnerApi.Pipeline pipeline, Struct options, @Nullable String retrievalToken)
+      throws IOException {
+    return invokeWithExecutor(pipeline, options, retrievalToken, this.executorService);
+  }
+
+  private ListeningExecutorService createExecutorService(String name) {
+    ThreadFactory threadFactory =
+        new ThreadFactoryBuilder().setNameFormat(name).setDaemon(true).build();
+    return MoreExecutors.listeningDecorator(Executors.newCachedThreadPool(threadFactory));
+  }
+
+  protected JobInvoker(String name) {
+    this.executorService = createExecutorService(name);
+  }
 }
